modified phone numbre data type

+++ GasotecFactureCreator/Controller/SalesCheckerController.cs
@@ -6,7 +6,7 @@ public static class SalesCheckerController
 {
     public static SalesChecker CurrentSalesChecker { get; private set; }
 
-    public static void CreateSalesChecker(string name, string address, int phoneNumber, string email, string Nit)
+    public static void CreateSalesChecker(string name, string address, long phoneNumber, string email, string Nit)
     {
         CurrentSalesChecker = new SalesChecker(name, address, phoneNumber, email, Nit);
     }
